PersonRepository Ref #4 #6

+++ src/main/java/model/PersonRepository.java
@@ -8,13 +8,26 @@ public class PersonRepository {
     private ArrayList<Person> personRepository;
     private int socialNumber;
 
-    public PersonRepository (ArrayList<Person> personRepository) {
+    //Constructors
+    public static PersonRepository createPersonRepository(ArrayList<Person> personRepository){
+        return new PersonRepository(personRepository);
+    }
+
+    private PersonRepository (ArrayList<Person> personRepository) {
         this.personRepository = personRepository;
     }
 
-    public void addPersonToPersonRepository(Person person){
-        if(!personRepository.contains(person)){
-            personRepository.add(person);
+    //Methods
+    public Person addPersonToPersonRepository(Person person) {
+        if (person == null) {
+            throw new NullPointerException("Person can't be null");
+        } else {
+            if (personRepository.contains(person)) {
+                throw new IllegalArgumentException("Person already exists in PersonRepository");
+            } else {
+                personRepository.add(person);
+            }
+            return person;
         }
     }
 
@@ -32,25 +45,30 @@ public class PersonRepository {
         return null;
     }
 
-    public void removePersonFromPersonRepository(Person person){
-        if(personRepository.contains(person)){
-            personRepository.remove(person);
-        } else{
-            throw new RuntimeException("" +
-                    "-------------------------------" +
-                    "Can't remove. Person does not belong." +
-                    "-------------------------------");
+    public Person removePersonFromPersonRepository(Person person){
+        if (person == null) {
+            throw new NullPointerException("Person can't be null");
+        } else {
+            if(personRepository.contains(person)){
+                personRepository.remove(person);
+            } else{
+                throw new RuntimeException("Can't remove. Person does not belong.");
+            }
         }
+        return person;
     }
 
+    //Gets
     public ArrayList<Person> getPersonRepository() {
         return personRepository;
     }
 
+    //Sets
     public void setPersonRepository(ArrayList<Person> personRepository) {
         this.personRepository = personRepository;
     }
 
+    //Override
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
